LTX (experimental): cast cache fingerprint sample to float32 before numpy
numpy has no native bfloat16; .numpy() on a bf16 tensor raises
TypeError. The text contexts + PE tensors are bf16 in production.

Cast the 8-element sample to float32 before .numpy().tobytes() — bit-
lossless for fingerprinting (float32 strictly contains bf16) and the
cost is meaningless (8 elements). Fix for the TypeError that surfaced
on the first call of the new session.

+++ server/payload_ltx.py
@@ -158,7 +158,13 @@ def compute_constants_session_id(
                 min(n // 3, n - 1),
                 min((2 * n) // 5, n - 1),
             })
-            sample = flat[positions].contiguous().cpu().numpy().tobytes()
+            # Cast to float32 before .numpy() — numpy lacks a native
+            # bfloat16 dtype (PyTorch's text-context and PE tensors are
+            # bf16). Lossless for fingerprinting since float32 has
+            # strictly more precision than bf16; we're hashing bits, not
+            # values, so any bit-distinct bf16 samples remain
+            # bit-distinct after the cast.
+            sample = flat[positions].to(torch.float32).contiguous().cpu().numpy().tobytes()
             h.update(sample)
     # Constant-relevant meta: pe_split_mode (per-key bool), flags
     # (transformer_options run_vx etc), and the subset of none_keys that
